pyproject.toml
[tool.poetry]
name = "depth-anything-v2"
version = "0.1.0"
description = ""
authors = ["Leo Innocenzi <leo.innocenzi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.4.0"
onnx = "^1.16.2"
onnxconverter-common = "^1.14.0"
opencv-python = "^4.10.0.84"
onnxruntime = "^1.18.1"
psutil = "^6.0.0"
torchvision = "^0.19.0"
pyyaml = "^6.0.1"
plotly = "^5.23.0"
matplotlib = "^3.9.1.post1"
tqdm = "^4.66.5"
coremltools = "^7.2"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
